+++ src/OpenTelemetry/CHANGELOG.md
@@ -15,8 +15,24 @@
   ([#2657](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2657))
 
 * Remove MetricStreamConfiguration.Aggregation, as the feature to customize
-aggregation is not implemented yet.
-([#2660](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2660))
+  aggregation is not implemented yet.
+  ([#2660](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2660))
+
+* Removed the public property `HistogramMeasurements` and added a public method
+  `GetHistogramBuckets` instead. Renamed the class `HistogramMeasurements` to
+  `HistogramBuckets` and added an enumerator of type `HistogramBucket` for
+  enumerating `BucketCounts` and `ExplicitBounds`. Removed `GetBucketCounts` and
+  `GetExplicitBounds` methods from `MetricPoint`.
+  ([#2664](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2664))
+
+* Refactored temporality setting to align with the latest spec.
+  ([#2666](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2666))
+
+* Removed the public properties `LongValue`, `DoubleValue`, `StartTime`, and
+  `EndTime` in favor of their counterpart public methods `GetCounterSumLong`,
+  `GetCounterSumDouble`, `GetGaugeLastValueLong`, `GetGaugeLastValueDouble`,
+  `GetStartTime`, and `GetEndTime`.
+  ([#2667](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2667))
 
 ## 1.2.0-beta2